Made hamming return type int to account for '-1' error case.

+++ exercises/hamming/src/example.c
@@ -2,13 +2,13 @@
 #include <string.h>
 #include "hamming.h"
 
-size_t compute(const char *lhs, const char *rhs)
+int compute(const char *lhs, const char *rhs)
 {
-   size_t count = -1;
+   int count = -1;
    if (lhs && rhs && (strlen(lhs) == strlen(rhs))) {
       count = 0;
-      size_t length = strlen(lhs);
-      for (size_t i = 0; i < length; ++i) {
+      int length = (int)strlen(lhs);
+      for (int i = 0; i < length; ++i) {
          if (lhs[i] != rhs[i]) {
             ++count;
          }
